requirements.txt
PyQt6>=6.6.0,<7.0.0
torch>=2.0.0
torchvision>=0.15.0
torchxrayvision>=1.3.0
grad-cam>=1.5.0
opencv-python-headless>=4.8.0
Pillow>=10.0.0
pydicom>=2.4.0
numpy>=1.24.0
reportlab>=4.0.0,<5.0.0
